pyproject.toml
[project]
authors = [
  {name = "Harsh Dadhich", email = "hdadhich01@gmail.com"}
]
dependencies = [
  "beautifulsoup4",
  "deta",
  "faust-cchardet",
  "flask",
  "flask-compress",
  "flask-cors",
  "flask-limiter",
  "gunicorn",
  "httpx",
  "humanize",
  "lxml",
  "orjson",
  "portolan",
  "python-dotenv",
  "round-nutrition",
  "schedule",
  "sentry-sdk[flask]",
  "thefuzz[speedup]"
]
description = "Default template for PDM package"
license = {text = "MIT"}
name = "api"
readme = "README.md"
requires-python = "==3.10.*"
version = "0.1.0"

[tool.pdm]
package-type = "application"
